infrastructure/inference-with-sagemaker-batch-transform/requirements.txt
boto3==1.36.3
sagemaker==2.219.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
transformers==4.48.1
torch==2.4.1
torchvision==0.19.1
safetensors==0.5.2
tokenizers==0.21.0
scikit-learn==1.5.2
Flask==3.1.0
gunicorn==22.0.0